+++ packages/workspace-fusion/src/lib/utils/grid/configureHighlightSelection.ts
@@ -1,0 +1,9 @@
+import { GridController } from '@workspace/grid';
+import { FusionMediator } from '../../types';
+
+export function configureHighlightSelection<TData, TError>(
+	gridController: GridController<TData>,
+	mediator: FusionMediator<TData, TError>
+) {
+	mediator.selection.onSelectionChanged((val) => gridController.selectedNodes.setValue(val.map(({ id }) => id)));
+}
