+++ packages/workspace-fusion/src/lib/utils/grid/snapShotGridState.ts
@@ -1,0 +1,9 @@
+import { GridController } from '@workspace/grid';
+import { GridBookmark } from '../../types';
+
+export function snapshotGridState<TData>(gridController: GridController<TData>): GridBookmark {
+	return {
+		selectedNodes: gridController.selectedNodes.value,
+		columnState: gridController.columnState,
+	};
+}
